+++ src/plugin.ts
@@ -7,10 +7,11 @@ import { hasScript, loadScript } from "./utils";
  * @see [developers.google.com/tag-manager/devguide](https://developers.google.com/tag-manager/devguide)
  */
 export interface DataLayerObject extends Record<string, any> {
-  event: string;
+  event?: string;
 }
 
 declare global {
+  // eslint-disable-next-line jsdoc/require-jsdoc
   interface Window {
     /**
      * `dataLayer` used by GTM.